+++ interpreter/byterun/pyvm2.py
@@ -3,8 +3,78 @@
 # 1. pyvm2 by Paul Swartz (z3p), from http://www.twistedmatrix.com/users/z3p/
 # 2. byterun by Ned Batchelder, github.com/nedbat/byterun
 
-import dis, inspect, operator, sys
-from byterun.pyobj import Frame, Block, Function
+import dis, operator, sys, collections, inspect, types
+
+class Frame(object):
+    def __init__(self, f_code, f_globals, f_locals, f_back):
+        self.f_code = f_code
+        self.f_globals = f_globals
+        self.f_locals = f_locals
+        self.f_back = f_back
+        self.stack = []
+        if f_back:
+            self.f_builtins = f_back.f_builtins
+        else:
+            self.f_builtins = f_locals['__builtins__']
+            if hasattr(self.f_builtins, '__dict__'):
+                self.f_builtins = self.f_builtins.__dict__
+
+        self.f_lineno = f_code.co_firstlineno
+        self.f_lasti = 0
+        self.block_stack = []
+
+Block = collections.namedtuple("Block", "type, handler, level")
+
+class Function(object):
+    __slots__ = [
+        'func_code', 'func_name', 'func_defaults', 'func_globals',
+        'func_locals', 'func_dict', 'func_closure',
+        '__name__', '__dict__', '__doc__',
+        '_vm', '_func',
+    ]
+
+    def __init__(self, name, code, globs, defaults, closure, vm):
+        self._vm = vm
+        self.func_code = code
+        self.func_name = self.__name__ = name or code.co_name
+        self.func_defaults = tuple(defaults)
+        self.func_globals = globs
+        self.func_locals = self._vm.frame.f_locals
+        self.__dict__ = {}
+        self.func_closure = closure
+        self.__doc__ = code.co_consts[0] if code.co_consts else None
+
+        # Sometimes, we need a real Python function.  This is for that.
+        kw = {
+            'argdefs': self.func_defaults,
+        }
+        if closure:
+            kw['closure'] = tuple(make_cell(0) for _ in closure)
+        self._func = types.FunctionType(code, globs, **kw)
+
+    def __get__(self, instance, owner):
+        if instance is not None:
+            return Method(instance, owner, self)
+        return self
+
+    def __call__(self, *args, **kwargs):
+        callargs = inspect.getcallargs(self._func, *args, **kwargs)
+        frame = self._vm.make_frame(
+            self.func_code, callargs, self.func_globals, {}
+        )
+        return self._vm.run_frame(frame)
+
+class Method(object):
+    def __init__(self, obj, _class, func):
+        self.im_self = obj
+        self.im_class = _class
+        self.im_func = func
+
+    def __call__(self, *args, **kwargs):
+        if self.im_self is not None:
+            return self.im_func(self.im_self, *args, **kwargs)
+        else:
+            return self.im_func(*args, **kwargs)
 
 
 class VirtualMachineError(Exception):
@@ -510,3 +580,8 @@ class VirtualMachine(object):
 
     def byte_STORE_LOCALS(self):
         self.frame.f_locals = self.pop()
+
+def make_cell(value):
+    # Thanks to Alex Gaynor for help with this bit of twistiness.
+    fn = (lambda x: lambda: x)(value)
+    return fn.__closure__[0]
